requirements.txt
discord.py==2.3.1
python-dotenv==1.0.1
Requests==2.31.0
